requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
cloudscraper>=1.2.71
lxml>=4.9.0
PyYAML>=6.0
click>=8.1.0

# Google Cloud
google-cloud-storage>=2.10.0
google-cloud-compute>=1.14.0
google-cloud-dns>=0.35.1
google-cloud-build>=3.17.0
google-cloud-artifact-registry>=1.12.0
google-cloud-run>=0.6.0
google-cloud-functions>=1.11.1
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0

# Web scraping and processing
selenium>=4.15.0
webdriver-manager>=4.0.0
pillow>=10.0.0
cssutils>=2.7.0

# Form handling
flask>=2.3.0
flask-cors>=4.0.0
sendgrid>=6.10.0
dotmap>=1.3.30

# Logging and monitoring
loguru>=0.7.0
rich>=13.6.0

# Development
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0

# Added from the code block
PyYAML==6.0.1
requests==2.32.3
beautifulsoup4==4.12.3
loguru==0.7.2
dotmap==1.3.30
google-api-python-client==2.142.0
google-auth-httplib2==0.2.0 